CIRCSTORE-425 Refactor processor structure
